Update the version and bundle ID.
Change the marketing version from "3.3.1" to "3.3.2-fork".  Change the
build version from empty to "3.3.2".  Set the `CFBundleVersion` key in
the `ManDrake-Info.plist` file to `$(CURRENT_PROJECT_VERSION)` instead
of "70" so it automatically picks up the build version.

Change the root prefix of the bundle IDs from `org.sveinbjorn` to
`cc.donm`.

Update the `Credits.rtf` file that holds the About box text, adding a
link to this GitHub fork.